Rerun command works now.
This commit makes me want to rethink the idea of TestSuite. It almost seems
that there are now two, almost three distinct types of 'TestSuite'.

The first type is what I would call simply a container. It contains other
suites and tests just like a list. This is used for my toplevel suite and
that's about it.

The second I would call a 'supertest' something is made up of multiple
tests/suites which are completely self_contained when combined into this
testsuite. We see this version for generated tests for gem5.

Finally the third is something in between the two. It's what I would call
a module. When loading a module there might be many individual tests floating
around that rely on shared data but are not explicitly grouped into a test
suite. This makes it impossible to group these into a self_contained suite, but
we also can gauruntee that the entire module is self contained. The reason we
don't then just group these into a module self_contained suite is because there
might also be other self_contained suites in that module, we would have
a self_contained module holding a second self_contained suite.